Cross Platform FTW, Re-refactor, Hacking ZeroConf
It turns out that pychromecast is wrapping it's socket client in SSL, which somehow breaks it inside Plex. So, we'll just get rid of that for now.

Re-refactor package, as apparently, "Plex" is not an allowed name.

Hack the zeroconf file to include newer fixes, but backport bits that dropped Python 2 compatibility. Also drop the dependency on netifaces, and use the cross-platform library ifaddr instead.