Make CustomContainer work more like actual Plex containers
Not sure if it matters, but the goal is to create custom info inside the web UI that matches the proper Plex schema, and another set of endpoints that return data however I need it. Now we can define restrictions on children and attributes if so desired, but still not throw a full error that breaks the process.  :D